+++ inst/www/shared/shiny.js
@@ -4833,7 +4833,7 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
       var evt = jQuery.Event("shiny:inputchanged");
       evt.name = this.id;
       evt.value = fileInfo;
-      evt.inputType = 'fileupload';
+      evt.inputType = 'shiny.fileupload';
       $(document).trigger(evt);
 
       this.makeRequest('uploadEnd', [this.jobId, this.id], function (response) {
@@ -4844,12 +4844,6 @@ var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol
         self.onError(error);
       });
       this.$bar().text('Finishing upload');
-
-      // Trigger event when all files are finished uploading.
-      var evt = jQuery.Event("shiny:fileuploaded");
-      evt.name = this.id;
-      evt.files = fileInfo;
-      $(document).trigger(evt);
     };
     this.onError = function (message) {
       this.$setError(message || '');