+++ NEWS.md
@@ -24,7 +24,7 @@ in shiny apps. For more info, see the documentation (`?updateQueryString` and `?
 
 ### Bug fixes
 
-* Fixed ([#1511](https://github.com/rstudio/shiny/issues/1511)): `fileInput`s did not trigger the `shiny:inputchanged` event on the client. ([#1541](https://github.com/rstudio/shiny/pull/1541))
+* Fixed ([#1511](https://github.com/rstudio/shiny/issues/1511)): `fileInput`s did not trigger the `shiny:inputchanged` event on the client. Also removed `shiny:fileuploaded` JavaScript event, because it is no longer needed after this fix. ([#1541](https://github.com/rstudio/shiny/pull/1541), [#1570](https://github.com/rstudio/shiny/pull/1570))
 
 * Fixed ([#1472](https://github.com/rstudio/shiny/issues/1472)): With a Progress object, calling `set(value=NULL)` made the progress bar go to 100%. Now it does not change the value of the progress bar. The documentation also incorrectly said that setting the `value` to `NULL` would hide the progress bar. ([#1547](https://github.com/rstudio/shiny/pull/1547))